Chess UI update
